backend/requirements.txt
bcrypt==3.1.6
dnspython==1.16.0
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.17.0
Flask-PyMongo==2.2.0
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
PyJWT==1.7.1
pylint==2.2.2
pymacaroons==0.13.0
pymongo==3.7.2
requests==2.21.0
six==1.12.0

